# openapi.yaml
openapi: 3.1.0
info:
  title: Puntazo API · programa gratuito
  version: 0.3.0-contract
  description: |
    Contrato de lanzamiento gratuito para programas Sellos y Puntos.
    Cada operación marcada IMPLEMENTED cuenta con código, persistencia, autorización
    y pruebas de contrato/integración dentro del repositorio aislado de preview.
  x-puntazo-slice: FREE-LAUNCH-SELLOS-PUNTOS
  x-puntazo-authority: ACORDADO_PARA_DEMO
  x-puntazo-approved-decisions: [PC-04, PC-09, PC-10, PC-11, PC-13, PC-14]
  x-puntazo-partial-decisions: [PC-12]
  x-puntazo-deferred: [PC-01, PC-03, PC-05, PC-06, PC-07, PC-08]
servers:
  - url: https://puntazo-preview.blakyta3d.duckdns.org/api/v1
    description: Staging VPS; el gateway web es el único servicio público
  - url: http://localhost:8080/v1
    description: Desarrollo local de API
security:
  - bearerAuth: []
...
paths:
...
  /backoffice/campaigns:
    get:
      tags: [Backoffice]
      operationId: listReferralCampaigns
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      responses:
        '200': { description: Hasta 200 campañas ordenadas por ID descendente, en data }
    post:
      tags: [Backoffice]
      operationId: createReferralCampaign
      summary: Requiere ADMIN_SISTEMA; rechaza ventanas superpuestas del mismo programa
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      parameters: [{ $ref: '#/components/parameters/BackofficeRequest' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReferralCampaignInput' }
      responses:
        '201': { description: Campaña creada y códigos de comercios existentes generados }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Ventana superpuesta }
  /backoffice/campaigns/{id}/active:
    patch:
      tags: [Backoffice]
      operationId: setReferralCampaignActive
      summary: Requiere ADMIN_SISTEMA; no altera atribuciones previas
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
        - { $ref: '#/components/parameters/BackofficeRequest' }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [active], properties: { active: { type: boolean } } } } }
      responses:
        '204': { description: Estado guardado y auditado }
  /backoffice/influencers:
    get:
      tags: [Backoffice]
      operationId: listReferralInfluencers
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      responses:
        '200': { description: Hasta 500 influencers, en data }
    post:
      tags: [Backoffice]
      operationId: createReferralInfluencer
      summary: Requiere ADMIN_SISTEMA
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      parameters: [{ $ref: '#/components/parameters/BackofficeRequest' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name, contact], properties: { name: { type: string }, contact: { type: string } } } } }
      responses:
        '201': { description: Influencer creado y auditado }
  /backoffice/codes:
    get:
      tags: [Backoffice]
      operationId: listReferralCodes
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      responses:
        '200': { description: Hasta 1000 códigos, en data }
    post:
      tags: [Backoffice]
      operationId: createInfluencerReferralCode
      summary: Requiere ADMIN_SISTEMA; código único normalizado a mayúsculas
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      parameters: [{ $ref: '#/components/parameters/BackofficeRequest' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [code, campaign_id, influencer_id], properties: { code: { type: string }, campaign_id: { type: integer }, influencer_id: { type: integer } } } } }
      responses:
        '201': { description: Código creado y auditado }
  /backoffice/codes/{id}/active:
    patch:
      tags: [Backoffice]
      operationId: setReferralCodeActive
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
        - { $ref: '#/components/parameters/BackofficeRequest' }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [active], properties: { active: { type: boolean } } } } }
      responses:
        '204': { description: Estado guardado y auditado }
  /backoffice/attributions:
    get:
      tags: [Backoffice]
      operationId: listReferralAttributions
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      responses:
        '200': { description: Hasta 1000 marcas atribuidas con condiciones congeladas, en data }
  /backoffice/rewards:
    get:
      tags: [Backoffice]
      operationId: listReferralRewards
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      responses:
        '200': { description: Hasta 1000 recompensas nacidas de cobros verificados, en data }
...
  /backoffice/brands/{id}/credit-allocations:
    get:
      tags: [Backoffice]
      operationId: getMerchantCreditAccount
      summary: Requiere FINANZAS; saldo y ajustes manuales de una marca referente
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
      responses:
        '200':
          description: Saldo disponible y hasta 200 asignaciones auditadas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MerchantCreditAccountEnvelope' }
    post:
      tags: [Backoffice]
      operationId: recordManualMerchantCredit
      summary: Requiere FINANZAS; registra un reembolso externo ya realizado contra una factura pagada
      description: La referencia externa es declarada por Finanzas y no verifica por sí misma una transferencia. No modifica el importe recurrente de Mercado Pago.
      security: [{ backofficeCookie: [] }]
      x-puntazo-implementation: IMPLEMENTED
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
        - { $ref: '#/components/parameters/BackofficeRequest' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MerchantCreditAllocationInput' }
      responses:
        '201':
          description: Asignación registrada con idempotencia y auditoría
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MerchantCreditAllocationEnvelope' }
        '409': { description: Saldo insuficiente, factura agotada o referencia duplicada }
...
components:
  securitySchemes:
    backofficeCookie:
      type: apiKey
      in: cookie
      name: puntazo_backoffice
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    BackofficeRequest:
      name: X-Backoffice-Request
      in: header
      required: true
      schema: { type: string, const: '1' }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: El actor no posee el tipo de cuenta, marca o sucursal requerida
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
    MerchantCreditAllocationInput:
      type: object
      required: [provider_invoice_id, amount_minor, external_settlement_reference]
      properties:
        provider_invoice_id: { type: string, minLength: 1 }
        amount_minor: { type: integer, format: int64, minimum: 1, description: Centavos de ARS }
        external_settlement_reference: { type: string, minLength: 6, maxLength: 120 }
    MerchantCreditAllocation:
      type: object
      required: [id, brand_id, provider_invoice_id, provider_payment_id, amount_minor, currency, external_settlement_reference, status, created_at]
      properties:
        id: { type: integer, format: int64 }
        brand_id: { type: integer, format: int64 }
        provider_invoice_id: { type: string }
        provider_payment_id: { type: string }
        amount_minor: { type: integer, format: int64 }
        currency: { type: string, enum: [ARS] }
        external_settlement_reference: { type: string }
        status: { type: string, enum: [RECORDED, RECOVERY_DUE] }
        recovery_due_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    MerchantCreditAllocationEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/MerchantCreditAllocation' }
    MerchantCreditAccountEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [balance_minor, allocations]
          properties:
            balance_minor: { type: integer, format: int64, description: Centavos de ARS; puede ser negativo tras una reversión }
            allocations:
              type: array
              items: { $ref: '#/components/schemas/MerchantCreditAllocation' }
    ReferralCampaignInput:
      type: object
      required: [name, program_type, discount_bps, discount_charges, reward_bps, reward_charges, starts_at, ends_at]
      properties:
        name: { type: string, minLength: 3, maxLength: 120 }
        program_type: { type: string, enum: [SELLOS, PUNTOS] }
        discount_bps: { type: integer, minimum: 0, maximum: 10000 }
        discount_charges: { type: integer, minimum: 0, maximum: 36 }
        reward_bps: { type: integer, minimum: 0, maximum: 10000 }
        reward_charges: { type: integer, minimum: 0, maximum: 36 }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
...
    RequestId:
      type: string
      format: uuid
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [INVALID_REQUEST, ACCOUNT_TYPE_REQUIRED, IDENTITY_TOKEN_INVALID, INVITATION_INVALID, INVITATION_EMAIL_MISMATCH, SELF_ROLE_CHANGE_FORBIDDEN, EMAIL_VERIFICATION_REQUIRED, UNAUTHENTICATED, RECENT_AUTH_REQUIRED, FORBIDDEN, NOT_FOUND, EMAIL_EXISTS, RESOURCE_CONFLICT, PROGRAM_TYPE_HAS_BENEFITS, PROGRAM_TYPE_IMMUTABLE, IDEMPOTENCY_CONFLICT, IDEMPOTENCY_IN_PROGRESS, PREVIEW_EXPIRED, PREVIEW_CONSUMED, PREVIEW_CHANGED, PRECONDITION_FAILED, OWNERSHIP_TRANSFER_REQUIRED, INSUFFICIENT_BALANCE, DEMO_SIGNUP_DISABLED, DEMO_ACCESS_DENIED, MEDIA_TOO_LARGE, MEDIA_TYPE_UNSUPPORTED, MEDIA_STORAGE_UNAVAILABLE, RATE_LIMITED, DEPENDENCY_UNAVAILABLE, INTERNAL_ERROR]
            message: { type: string }
            details: { type: object, additionalProperties: true }
        request_id: { $ref: '#/components/schemas/RequestId' }
...